+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Next
 
+* Updated Go dependency to go1.24: needed for the safer `runtime.AddCleanUp`.
 * Fixed the issue where Client objects never got garbage collected, due to cycles in finalizers.
   * Internal refactoring changing `runtime.SetFinalizer` to `runtime.AddCleanUp`.
 * Added GOPJRT_TEXT_STABLE_HLO option to force XlaBuilder to use StableHlo text representation.