+++ docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Gopjrt Changelog
 
+# Next
+
+* Updated Go dependency to go1.24: needed for the safer `runtime.AddCleanUp`.
+* Fixed the issue where Client objects never got garbage collected, due to cycles in finalizers.
+  * Internal refactoring changing `runtime.SetFinalizer` to `runtime.AddCleanUp`.
+* Added GOPJRT_TEXT_STABLE_HLO option to force XlaBuilder to use StableHlo text representation.
+  * Use with `-vmodule=compile=2` to enable logging of each program compiled. 
+
 # v0.7.0 - 2024/04/30
 
 * Renamed Gather() parameter offsetAxes to offsetOutputAxes to avoid confusion.