+++ docs/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 # Next
 
-* Updated Go dependency to go1.24: needed for the safer `runtime.AddCleanUp`.
+* Updated Go dependency to go1.24: needed for the safer `runtime.AddCleanUp` used in `pjrt` now.
+* Moved internal packages under a `internal` subdirectory, including code generating commands.
 * Fixed the issue where Client objects never got garbage collected, due to cycles in finalizers.
   * Internal refactoring changing `runtime.SetFinalizer` to `runtime.AddCleanUp`.
 * Added GOPJRT_TEXT_STABLE_HLO option to force XlaBuilder to use StableHlo text representation.